# openapi.yaml
openapi: 3.1.0
info:
  title: AITP Control Plane API
  version: 0.1.0
  description: |
    Registry, audit log, revocation list, and webhook fan-out for an AITP
    (Agent Identity Trust Protocol) deployment. See README.md and
    docs/api.md for the prose reference.
  license:
    name: See LICENSE
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: https://control-plane.example.com
    description: Production placeholder
...
security:
  - apiKey: []
  - {}
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: API key from `API_KEYS` env allowlist. Required for gated routes; unset in dev disables auth.
    enrollmentToken:
      type: http
      scheme: bearer
      description: One-time bearer issued by `POST /api/registry/enroll`. Used only on `POST /api/registry/agents`.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
      description: Opaque client-chosen key. Replaying the same key returns the cached response without re-running the handler.
    namespace:
      name: namespace
      in: query
      required: false
      schema:
        type: string
      description: Tenant scope filter. Omitting it returns rows across all namespaces.
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
    Agent:
      type: object
      required: [aid, displayName, handshakeEndpoint, offeredCaps, status]
      properties:
        aid: { type: string, example: "aid:pubkey:z:abc..." }
        displayName: { type: string }
        handshakeEndpoint: { type: string, format: uri }
        offeredCaps:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [active, expired, deregistered]
        namespace: { type: string, default: default }
        org: { type: string, nullable: true }
        cloud: { type: string, nullable: true }
        registeredAt: { type: string, format: date-time }
        lastEnrolledAt: { type: string, format: date-time, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        manifestUrl: { type: string }
        agentManifestHint: { type: string, nullable: true }
        manifestJson:
          type: string
          description: Inline ManifestEnvelope JSON. Present only when `?include_manifest=true`.
    EnrollResponse:
      type: object
      required: [token, expiresIn, aid]
      properties:
        token: { type: string }
        expiresIn: { type: integer, example: 300 }
        aid: { type: string }
...
paths:
...
  /api/registry/enroll:
    post:
      tags: [registry]
      summary: Issue a one-time enrollment token from a signed manifest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                manifest: { type: object }
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnrollResponse" }
        "400": { description: Invalid manifest, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: Rate limited }
  /api/registry/agents:
    get:
      tags: [registry]
      summary: Discover agents by capability or AID
      security: []
      parameters:
        - name: capability
          in: query
          schema: { type: string }
        - name: aid
          in: query
          schema: { type: string }
        - name: displayName
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/namespace"
        - name: include_manifest
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Matching agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
    post:
      tags: [registry]
      summary: Register an agent with a previously-issued enrollment token
      security:
        - enrollmentToken: []
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
        - name: X-Aitp-Namespace
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest]
              properties:
                manifest: { type: object }
      responses:
        "201":
          description: Registered
        "400": { description: Bad manifest }
        "401": { description: Bad token }
        "429": { description: Rate limited }
  /api/registry/agents/{aid}:
    parameters:
      - name: aid
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [registry]
      summary: Fetch one agent
      security: []
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Agent" } } } }
        "404": { description: Not found }
    delete:
      tags: [registry]
      summary: Deregister an agent (admin)
      responses:
        "204": { description: Deregistered }
        "404": { description: Not found }
  /api/registry/agents/{aid}/manifest:
    parameters:
      - name: aid
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [registry]
      summary: Fetch the cached ManifestEnvelope for an agent
      security: []
      responses:
        "200": { description: ManifestEnvelope JSON }
        "404": { description: Not found }
...